+++ packages/renderer/src/renderer/screens/Home/index.tsx
@@ -6,6 +6,7 @@ import { useEffect } from 'react';
 import { useAddresses } from '@app/contexts/main/Addresses';
 import { useAppSettings } from '@app/contexts/main/AppSettings';
 import { useBootstrapping } from '@app/contexts/main/Bootstrapping';
+import { useCogMenu } from '@app/contexts/main/CogMenu';
 import { useConnections } from '@app/contexts/common/Connections';
 import { useEvents } from '@app/contexts/main/Events';
 import { useInitIpcHandlers } from '@app/hooks/useInitIpcHandlers';
@@ -17,7 +18,6 @@ import { Footer } from './Footer';
 import { Manage } from './Manage';
 import { FixedFlexWrapper, IconWrapper } from './Wrappers';
 import IconSVG from '@app/svg/polkadotIcon.svg?react';
-import { Flip, toast } from 'react-toastify';
 import { version } from '../../../../package.json';
 
 /** Library */
@@ -30,7 +30,6 @@ import {
 } from '@polkadot-live/ui/components';
 import { ScrollWrapper } from '@polkadot-live/ui/styles';
 import { useSideNav } from '@polkadot-live/ui/contexts';
-import { useHelp } from '@app/contexts/common/Help';
 
 /** Types */
 import type { ChainID } from '@polkadot-live/types/chains';
@@ -54,22 +53,17 @@ export const Home = () => {
     silenceOsNotifications,
     handleDockedToggle,
     handleSideNavCollapse,
-    handleToggleSilenceOsNotifications,
   } = useAppSettings();
 
-  const {
-    appLoading,
-    isConnecting,
-    isAborting,
-    online: isOnline,
-    handleInitializeAppOnline,
-    handleInitializeAppOffline,
-    setIsAborting,
-    setIsConnecting,
-  } = useBootstrapping();
-
+  const { appLoading } = useBootstrapping();
   const { selectedId, setSelectedId } = useSideNav();
-  const { openHelp } = useHelp();
+  const {
+    getAppFlags,
+    getConnectionButtonText,
+    getMenuItems,
+    handleConnectClick,
+    handleSilenceNotifications,
+  } = useCogMenu();
 
   useEffect(() => {
     // Listen for event callbacks.
@@ -115,23 +109,6 @@ export const Home = () => {
    * TODO: Put in context.
    */
 
-  /// Menu connect button text.
-  const getConnectionButtonText = () => {
-    if (isConnecting || appLoading) {
-      return 'Abort';
-    } else if (isOnline) {
-      return 'Disconnect';
-    } else {
-      return 'Connect';
-    }
-  };
-
-  /// Handle abort connecting.
-  const handleAbortConnecting = () => {
-    setIsAborting(true);
-    ConfigRenderer.abortConnecting = true;
-  };
-
   return (
     <>
       <Header
@@ -152,110 +129,14 @@ export const Home = () => {
         onRestoreWindow={() => window.myAPI.restoreWindow('base')}
         version={version}
       >
-        {/* TODO: Lift menu functions to context */}
+        {/* Logic in cog menu context */}
         <Menu
-          menuItems={[
-            {
-              label: 'Accounts',
-              disabled: appLoading,
-              onClick: () => {
-                window.myAPI.openWindow('import');
-                window.myAPI.umamiEvent('window-open-accounts', null);
-              },
-            },
-            {
-              label: 'OpenGov',
-              disabled: appLoading,
-              onClick: () => {
-                window.myAPI.openWindow('openGov');
-                window.myAPI.umamiEvent('window-open-openGov', null);
-              },
-            },
-            {
-              label: 'Settings',
-              disabled: appLoading,
-              onClick: () => {
-                window.myAPI.openWindow('settings');
-                window.myAPI.umamiEvent('window-open-settings', null);
-              },
-            },
-            {
-              label: 'Exit',
-              disabled: false,
-              appendSeparator: true,
-              onClick: () => {
-                window.myAPI.quitApp();
-              },
-            },
-            {
-              label: 'Disclaimer',
-              disabled: false,
-              onClick: () => {
-                openHelp('help:docs:disclaimer');
-              },
-            },
-            {
-              label: 'Privacy',
-              disabled: false,
-              onClick: () => {
-                openHelp('help:docs:privacy');
-              },
-            },
-          ]}
-          appFlags={{
-            isConnecting,
-            isOnline,
-            isAborting,
-            isLoading: appLoading,
-          }}
+          appFlags={getAppFlags()}
+          menuItems={getMenuItems()}
           connectLabel={getConnectionButtonText()}
+          onConnectClick={handleConnectClick}
           silenceOsNotifications={silenceOsNotifications}
-          onSilenceNotifications={() => {
-            handleToggleSilenceOsNotifications();
-
-            ConfigRenderer.portToSettings?.postMessage({
-              task: 'settings:set:silenceOsNotifications',
-              data: {
-                silenced: !silenceOsNotifications,
-              },
-            });
-          }}
-          onConnectClick={async () => {
-            if (isConnecting || appLoading) {
-              handleAbortConnecting();
-            } else if (isOnline) {
-              await handleInitializeAppOffline();
-            } else {
-              // Confirm online connection.
-              const status: boolean =
-                (await window.myAPI.sendConnectionTaskAsync({
-                  action: 'connection:getStatus',
-                  data: null,
-                })) || false;
-
-              if (status) {
-                // Handle going online.
-                setIsConnecting(true);
-                await handleInitializeAppOnline();
-                setIsConnecting(false);
-              } else {
-                // Render error alert.
-                toast.error('You are offline.', {
-                  position: 'top-center',
-                  autoClose: 3000,
-                  hideProgressBar: true,
-                  closeOnClick: true,
-                  closeButton: false,
-                  pauseOnHover: false,
-                  draggable: false,
-                  progress: undefined,
-                  theme: 'dark',
-                  transition: Flip,
-                  toastId: 'toast-connection', // prevent duplicate alerts
-                });
-              }
-            }
-          }}
+          onSilenceNotifications={handleSilenceNotifications}
         />
       </Header>
